+++ packages/charts/src/chart_types/heatmap/state/selectors/is_brush_available.ts
@@ -7,17 +7,21 @@
  */
 
 import { createCustomCachedSelector } from '../../../../state/create_selector';
+import { getSettingsSpecSelector } from '../../../../state/selectors/get_settings_specs';
 import { getHeatmapConfigSelector } from './get_heatmap_config';
 
 /**
  * The brush is available only if a onBrushEnd listener is configured
  * @internal
  */
-export const isBrushAvailableSelector = createCustomCachedSelector([getHeatmapConfigSelector], (config): boolean => {
-  return Boolean(config.onBrushEnd) && config.brushTool.visible;
-});
+export const isBrushAvailableSelector = createCustomCachedSelector(
+  [getHeatmapConfigSelector, getSettingsSpecSelector],
+  (config, { onBrushEnd }): boolean => {
+    return Boolean(onBrushEnd) && config.brushTool.visible;
+  },
+);
 
 /** @internal */
-export const isBrushEndProvided = createCustomCachedSelector([getHeatmapConfigSelector], (config): boolean => {
-  return Boolean(config.onBrushEnd);
+export const isBrushEndProvided = createCustomCachedSelector([getSettingsSpecSelector], ({ onBrushEnd }): boolean => {
+  return Boolean(onBrushEnd);
 });