+++ packages/charts/src/components/brush/brush.tsx
@@ -27,7 +27,7 @@ interface StateProps {
   projectionContainer: Dimensions;
   isBrushing: boolean | undefined;
   isBrushAvailable: boolean | undefined;
-  brushArea: Dimensions | null;
+  brushEvent: Dimensions | null;
   zIndex: number;
 }
 
@@ -91,12 +91,13 @@ class BrushToolComponent extends React.Component<StateProps> {
   }
 
   private drawCanvas = () => {
-    const { brushArea, mainProjectionArea } = this.props;
+    const { brushEvent, mainProjectionArea } = this.props;
+
     const { ctx } = this;
-    if (!ctx || !brushArea) {
+    if (!ctx || !brushEvent) {
       return;
     }
-    const { top, left, width, height } = brushArea;
+    const { top, left, width, height } = brushEvent;
     withContext(ctx, () => {
       ctx.scale(this.devicePixelRatio, this.devicePixelRatio);
       withClip(
@@ -145,7 +146,7 @@ const mapStateToProps = (state: GlobalChartState): StateProps => {
       },
       isBrushing: false,
       isBrushAvailable: false,
-      brushArea: null,
+      brushEvent: null,
       zIndex: 0,
     };
   }
@@ -155,7 +156,7 @@ const mapStateToProps = (state: GlobalChartState): StateProps => {
     mainProjectionArea: getInternalMainProjectionAreaSelector(state),
     isBrushAvailable: getInternalIsBrushingAvailableSelector(state),
     isBrushing: getInternalIsBrushingSelector(state),
-    brushArea: getInternalBrushAreaSelector(state),
+    brushEvent: getInternalBrushAreaSelector(state),
     zIndex: state.zIndex,
   };
 };